// src/app/flows/cases/cases-create-casefile/components/cases-create-casefile-third-party/interfaces/cases-create-casefile-third-party-field-names.interface.ts
export interface ICasesCreateCasefileThirdPartyFieldNames {
  nameOrOrganisation: string;
  relationship: string;
  reference: string;
  addressLine1: string;
  addressLine2: string;
  addressLine3: string;
  addressLine4: string;
  addressLine5: string;
  postalOrZipCode: string;
  countryId: string;
}